refactor(point): Change amount and balanceAfter to Long type
- Integer 오버플로우 방지를 위해 Long 타입으로 변경
- amount: Integer → Long
- balanceAfter: Integer → Long
- 포인트 누적 시 Integer.MAX_VALUE(약 21억) 초과 가능성 대응

+++ src/main/java/com/golfRental/domain/point/dto/response/PointTransactionGetResponse.java
@@ -8,9 +8,9 @@ import java.time.LocalDateTime;
 @Builder
 public record PointTransactionGetResponse(
         Long transactionId,
-        Integer amount,
+        Long amount,
         PointTransactionType type,
-        Integer balanceAfter,
+        Long balanceAfter,
         LocalDateTime createdAt
 ) {
 }
